[FIX] .gitignore was ignoring custom js script
The custom js script 'move-build' was created previously to move the react build
into django. Adding CI exposed that this was not being pushed to GitHub due to
being erroneously being added to the .gitignore file.
NOTE: Build directories should still be ignored, as they will be created as part
of the Continuous Deployment pipeline.